Prompt Reverse Engineering/requirements.txt
torch
transformers
timm
Pillow
gradio
git+https://github.com/pharmapsychotic/clip-interrogator.git 